Devcontainer updates (GUI support, Nvidia GPU resources, Debug Probe now works properly, Pico TTY permissions fixed), testing and debugging continues.

+++ Source Code/pico_ws/mainboard_firmware/lib/Motor_lib/include/motor_control_lib/Motor.h
@@ -72,7 +72,7 @@ class Motor
         void set_pid_ctrl_speed(float rpm);
 
         // Set PWM output for blind PWM control.
-        void set_pwm_ctrl_speed(int pwm_out);
+        void set_pwm_ctrl_speed(uint16_t pwm_out);
 
         // Set motor direction.
         // The actual direction in which the motor is
